[FIX] TEMP FOR TESTING

+++ src/Engine/ECS/Components/Colliders/CircleCollider.cpp
@@ -1,7 +1,7 @@
 ﻿#include "pch.h"
 #include "CircleCollider.h"
 
-CircleCollider::CircleCollider(Entity* entity, int radius) : Collider2D(entity)
+CircleCollider::CircleCollider(Entity* entity, float radius) : Collider2D(entity)
 {
     mColliderType = ColliderType::CIRCLE;
     mRadius = radius;
